FRC-2022-NEW/offboard/quixpf/requirements.txt
matplotlib
numpy==1.24.1
Pillow
pyopengl==3.1.6
pyqtgraph==0.13.1
PyQt6==6.4.0
transformations
pynetworktables==2021.0.0
numpy-stl